feat(follow_button): Add logic for the following
When the follow button is pressed, the user from the profile page
is added to the friend list of the authenticated user. Unfortunately,
I could not get the integration with the authentication to work without
wrapping it in a try block. When I will merge with the proper authentication,
this will be easier.