fix(testing): Fix test classes related to follow button
1. In the firestore test class we will now remove all users when the test ends.
2. For testing the follow button, I changed the user search screen test class in order
to verify also if the follow button functionality works. Alice and Bob (the users between
which the interaction is tested) are already friends from Alice's point of view. Therefore,
when Alice presses the follow button, it should be in the unfollow state and then changes to
unfollow as Alice removes Bob as a friend.