feat: add clickable location to outfit post cards
Modified OutfitPostCard to make the PostLocation composable clickable by adding
an onClick callback parameter. The location text now triggers a callback when
tapped, allowing users to navigate to the map. Updated FeedScreen to pass the
location click handler through the component hierarchy from FeedScreen down to
FeedList and OutfitPostCard. This enables users to quickly view a post's
location on the map by tapping the location text below the post image.